Fixed unsigned long cast compiler warning
git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@2309 8e3573b8-cf2c-4f14-ef6d-137439e28b8b

+++ libcommon_c++/SRC/TC_Common/TC_SideIndex.cxx
@@ -148,7 +148,7 @@ void TC_SideIndex_c::ExtractString(
 	 if ( this->index_ != SIZE_MAX )
 	 {
    	    char szIndex[ TIO_FORMAT_STRING_LEN_VALUE ];
-            sprintf( szIndex, "%lu", this->index_ );
+            sprintf( szIndex, "%lu", static_cast<unsigned long>( this->index_ ));
 
             *psrSideIndex += " ";
             *psrSideIndex += szIndex;
